C++ cleanup
* Update C++ sample client to match Java
* Fix C++ handling of double
* C++ update for keep_alive and open_session
* C++ server crashing due to memory leak
* C++ set property by index

+++ remote/rtRemoteValueReader.cpp
@@ -89,7 +89,7 @@ rtRemoteValueReader::read(rtRemoteEnvironment* env, rtValue& to, rapidjson::Valu
     break;
     
     case RT_doubleType:
-    to.setFloat(static_cast<double>(val->value.GetDouble()));
+    to.setDouble(val->value.GetDouble());
     break;
 
     case RT_stringType:
